Moving dialects into specifics submodule to avoid dependency problems
