+++ uPortal-utils/uPortal-utils-core/build.gradle
@@ -7,6 +7,7 @@ ext {
 
 dependencies {
     compile project(':uPortal-concurrency')
+    compile project(':uPortal-hibernate:uPortal-hibernate4-dialects')
 
     compile "com.fasterxml.jackson.core:jackson-annotations:${jacksonVersion}"
     compile "com.fasterxml.jackson.core:jackson-core:${jacksonVersion}"
